Update to Paper's 1.21.4 hard fork
This repo is mostly just so that I can get my other forks onto 1.21.4. I deleted the useless packets patch as supposedly it causes issues in 1.21.4

This process was awful, but I probably did things wrong as I wanted to keep the patches as feature patches instead of file patches.

I'm definitely going to go with file patches for my other forks.